Added telephone voice, old radio and jet pilot using the updated filtering classes
git-svn-id: https://mary.opendfki.de/repos/trunk@638 953a6561-930b-0410-b2a6-db37d1b2ae63

+++ java/de/dfki/lt/signalproc/demo/VoiceModificationParameters.java
@@ -39,8 +39,8 @@ public class VoiceModificationParameters {
      * 
      */
     public VoiceModificationParameters() {
-        fs = 16000.0f;  //Set default sampling rate
+        fs = 16000;  //Set default sampling rate
     }
     
-    float fs; //Sampling rate in Hz
+    int fs; //Sampling rate in Hz
 }
